+++ kermit/k95/ckonet.c
@@ -4462,6 +4462,10 @@ netinit() {
                 debug(F100,"SSH not installed","",0) ;
             }
         }
+#else
+#ifdef SSHBUILTIN
+        ssh_initialise();
+#endif /* SSHBUILTIN */
 #endif /* SSH_DLL */
     }